requirements.txt
Flask==1.1.2
line-bot-sdk==1.18.0
psycopg2==2.8.6
Flask-SQLAlchemy==2.4.4
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.24.1
Flask-APScheduler==1.11.0
requests==2.24.0
Flask-Cors==3.0.10